Create README.md
"The objective of this project is to utilize artificial intelligence in Python to create a smart stock portfolio! We will use artificial intelligence in the stock market to obtain stock quotes using Python, perform data preprocessing in Python, among various other steps. We will guide you through the project step by step and show you the results, so you can understand what the project is about and where we are heading. This new project is going to be very interesting, especially if you are interested in investments!"